Nuevos dto y manejo de excepciones

+++ src/main/java/com/example/vuelos/controllers/AerolineaController.java
@@ -1,6 +1,8 @@
 package com.example.vuelos.controllers;
 
-import com.example.vuelos.dtos.AerolineaDTO;
+import com.example.vuelos.controllers.dtos.AerolineaDTO;
+import com.example.vuelos.controllers.dtos.AerolineaRequestDTO;
+import com.example.vuelos.exceptions.ResourceNotFound;
 import com.example.vuelos.services.AerolineaService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -23,25 +25,26 @@ public class AerolineaController {
 
     @GetMapping("/{id}")
     public ResponseEntity<AerolineaDTO> getAerolineaById(@PathVariable Long id) {
-        return aerolineaService.findById(id)
-                .map(ResponseEntity::ok)
-                .orElse(ResponseEntity.notFound().build());
+        return ResponseEntity.ok(aerolineaService.findById(id)
+                .orElseThrow(() -> new ResourceNotFound("Aerolínea no encontrada con id: " + id)));
     }
 
     @PostMapping
-    public ResponseEntity<AerolineaDTO> createAerolinea(@RequestBody AerolineaDTO aerolineaDTO) {
-        return ResponseEntity.ok(aerolineaService.create(aerolineaDTO));
+    public ResponseEntity<AerolineaDTO> createAerolinea(@RequestBody AerolineaRequestDTO aerolineaRequestDTO) {
+        return ResponseEntity.ok(aerolineaService.create(aerolineaRequestDTO));
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<AerolineaDTO> updateAerolinea(@RequestBody AerolineaDTO aerolineaToUpdateDTO, @PathVariable Long id) {
-        return aerolineaService.update(id, aerolineaToUpdateDTO)
-                .map(ResponseEntity::ok)
-                .orElse(ResponseEntity.notFound().build());
+    public ResponseEntity<AerolineaDTO> updateAerolinea(@RequestBody AerolineaRequestDTO aerolineaRequestDTO, @PathVariable Long id) {
+        return ResponseEntity.ok(aerolineaService.update(id, aerolineaRequestDTO)
+                .orElseThrow(() -> new ResourceNotFound("Aerolínea no encontrada con id: " + id)));
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteAerolinea(@PathVariable Long id) {
+        if (!aerolineaService.findById(id).isPresent()) {
+            throw new ResourceNotFound("Aerolínea no encontrada con id: " + id);
+        }
         aerolineaService.delete(id);
         return ResponseEntity.noContent().build();
     }
