+++ src/main/java/com/example/vuelos/controllers/AeropuertoController.java
@@ -1,6 +1,8 @@
 package com.example.vuelos.controllers;
 
-import com.example.vuelos.dtos.AeropuertoDTO;
+import com.example.vuelos.controllers.dtos.AeropuertoDTO;
+import com.example.vuelos.controllers.dtos.AeropuertoRequestDTO;
+import com.example.vuelos.exceptions.ResourceNotFound;
 import com.example.vuelos.services.AeropuertoService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -23,25 +25,26 @@ public class AeropuertoController {
 
     @GetMapping("/{id}")
     public ResponseEntity<AeropuertoDTO> getAeropuertoById(@PathVariable Long id) {
-        return aeropuertoService.findById(id)
-                .map(ResponseEntity::ok)
-                .orElse(ResponseEntity.notFound().build());
+        return ResponseEntity.ok(aeropuertoService.findById(id)
+                .orElseThrow(() -> new ResourceNotFound("Aeropuerto no encontrado con id: " + id)));
     }
 
     @PostMapping
-    public ResponseEntity<AeropuertoDTO> createAeropuerto(@RequestBody AeropuertoDTO aeropuertoDTO) {
-        return ResponseEntity.ok(aeropuertoService.create(aeropuertoDTO));
+    public ResponseEntity<AeropuertoDTO> createAeropuerto(@RequestBody AeropuertoRequestDTO aeropuertoRequestDTO) {
+        return ResponseEntity.ok(aeropuertoService.create(aeropuertoRequestDTO));
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<AeropuertoDTO> updateAeropuerto(@RequestBody AeropuertoDTO aeropuertoToUpdateDTO, @PathVariable Long id) {
-        return aeropuertoService.update(id, aeropuertoToUpdateDTO)
-                .map(ResponseEntity::ok)
-                .orElse(ResponseEntity.notFound().build());
+    public ResponseEntity<AeropuertoDTO> updateAeropuerto(@RequestBody AeropuertoRequestDTO aeropuertoRequestDTO, @PathVariable Long id) {
+        return ResponseEntity.ok(aeropuertoService.update(id, aeropuertoRequestDTO)
+                .orElseThrow(() -> new ResourceNotFound("Aeropuerto no encontrado con id: " + id)));
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteAeropuerto(@PathVariable Long id) {
+        if (!aeropuertoService.findById(id).isPresent()) {
+            throw new ResourceNotFound("Aeropuerto no encontrado con id: " + id);
+        }
         aeropuertoService.delete(id);
         return ResponseEntity.noContent().build();
     }
